MAGECLOUD-1401: Move log resolving to Logger class instead of deploy preprocess (#124)
